Updates dependencies and the default for one option
The 'connectionKey' option defaulted to `connection`, but `req.connection` already exists in express, which causes some issues in certain circumstances, although admittedly I do not completely understand them, however I do understand that it is really really bad for me to override existing keys on req. Overall, I'm not a big fan of adding things onto `req` anymore. Right now, the current pattern I like doing is using weakmaps with the `req` as the key, and the things I want to extend the `req` with as the value, but that's beyond the scope of this commit